fix: 修复 CI 构建问题
修复内容:
1. 创建 client/assets 目录避免复制错误
2. 修复 renderer.cpp 类型转换警告
   - world_to_screen 添加 static_cast<float>
   - 避免 double -> float narrowing conversion
3. Windows 禁用客户端和可视化 demo
   - Windows 缺少 X11 依赖
   - 只构建服务器和压力测试工具
4. pathfinding_demo 仅在 Linux/macOS 构建

构建矩阵:
- Linux x64: server + client + pathfinding_demo
- macOS arm64: server + client + pathfinding_demo
- Windows x64: server only (无 GUI)

+++ client/src/renderer.cpp
@@ -155,8 +155,8 @@ Vec2 Renderer::screen_to_world(Vector2 screen_pos) const {
 
 Vector2 Renderer::world_to_screen(Vec2 world_pos) const {
   return Vector2{
-    world_pos.x + camera_offset_.x,
-    world_pos.y + camera_offset_.y
+    static_cast<float>(world_pos.x + camera_offset_.x),
+    static_cast<float>(world_pos.y + camera_offset_.y)
   };
 }
